% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.activity.class.R
\name{activity.class}
\alias{activity.class}
\alias{act_naei_rural}
\alias{act_naei_urban}
\alias{act_naei_motorway}
\alias{act_naei_rum}
\title{activity.class}
\usage{
act_naei_rural(...)

act_naei_urban(...)

act_naei_motorway(...)

act_naei_rum(...)
}
\value{
These functions make activity class objects, which .
}
\description{
Functions to build activity objects for use in embrs emission
models.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and Beddows references...
}
