% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus.objects.R
\name{bus.objets}
\alias{bus.objets}
\alias{bus_ice}
\alias{bus_bev}
\alias{vehicle_workhorse}
\title{bus objects}
\usage{
bus_ice(veh.wt, n = 1, name = NULL, ...)

bus_bev(veh.wt, n = 1, name = NULL, ...)

vehicle_workhorse(name = NULL, x = NULL, ...)
}
\arguments{
\item{veh.wt}{(required numeric) weight of vehicle in kg.}

\item{n}{(numeric) number of vehicles, defualt 1.}

\item{name}{(character) name of the vehicle object.}

\item{...}{other arguments, currently ignored}

\item{x}{(for workhorse only) a list of args to using when building
an \strong{embrs} object.}
}
\description{
Bus objects for use in \strong{embrs} models.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and Beddows references...
}
