% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coach.objects.R
\name{coach.objects}
\alias{coach.objects}
\alias{coach_ice}
\alias{coach_bev}
\title{coach objects}
\usage{
coach_ice(
  veh.wt = NULL,
  eng.fuel = NULL,
  euro.class = NULL,
  ...,
  veh.type = "coach"
)

coach_bev(veh.wt = NULL, ..., veh.type = "coach")
}
\arguments{
\item{veh.wt}{(required numeric) weight of vehicle in kg.}

\item{eng.fuel}{(required character) the engine fuel used by the vehicle,
e.g. diesel.}

\item{euro.class}{(required character) vehicle EURO classification,
e.g. PRE, I, II, III, IV, V or VI.}

\item{...}{other arguments, passed on to common vehicle build functions.
Arguments depend on the vehicle/engine/fuel combination. See embrs.vehicles
help for further details.}

\item{veh.type}{(character) type of vehicle, default coach.}
}
\value{
These functions make coach models for \strong{embrs} emissions
models. The main models coach_ice and coach_bev are internal compression engine
and battery electric buses, respectively. These functions extend bus method
used in Tivey et al (2023) to coaches.
}
\description{
Coach objects for use in \strong{embrs} models.
}
\note{
Work in progress...
}
\references{
Tivey, J., Davies, H.C., Levine, J.G., Zietsman, J., Bartington, S.,
Ibarra-Espinosa, S. and Ropkins, K, 2023. Meta-Analysis as Early Evidence on
the Particulate Emissions Impact of EURO VI on Battery Electric Bus Fleet
Transitions. Sustainability 15, 1522. \url{https://doi.org/10.3390/su15021522}
}
