% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_beddows.R
\name{ef_beddows}
\alias{ef_beddows}
\alias{ef_beddows_brake_pm}
\alias{ef_pm_brake_beddows}
\alias{ef_beddows_tyre_pm}
\alias{ef_beddows_road_pm}
\alias{ef_beddows_resusp_pm}
\title{Beddows Non-Exhaust Emissions Factors}
\usage{
ef_beddows_brake_pm(wt, output = "df")

ef_beddows_tyre_pm(wt, output = "df")

ef_beddows_road_pm(wt, output = "df")

ef_beddows_resusp_pm(wt, output = "df")
}
\arguments{
\item{wt}{vehicle weight (in kg)}

\item{output}{ignore for now...}
}
\description{
functions to estimated vehicle weight-based non-exhaust
PM2.5 and PM10 emission factors based on methods of Beddows.
}
\note{
These may be moving to vein at some point...
}
\references{
add Beddows and Harrison ref...
}
