% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_embrs.R
\name{ef_embrs}
\alias{ef_embrs}
\alias{ef_embrs1_nee_pm}
\alias{ef_embrs1_brake_pm2.5}
\alias{ef_embrs1_tyre_pm2.5}
\alias{ef_embrs1_road_pm2.5}
\alias{ef_embrs1_resusp_pm2.5}
\alias{ef_embrs1_brake_pm10}
\alias{ef_embrs1_tyre_pm10}
\alias{ef_embrs_road1_pm10}
\alias{ef_embrs1_resusp_pm10}
\alias{ef_embrs2_brake_pm2.5}
\alias{ef_embrs2_tyre_pm2.5}
\alias{ef_embrs2_brake_pm10}
\alias{ef_embrs1_road_pm10}
\alias{ef_embrs2_brake_pm}
\alias{ef_embrs2_tyre_pm}
\alias{ef_embrs2_tyre_pm10}
\title{embrs Non-Exhaust Emissions Factors}
\usage{
ef_embrs1_nee_pm(
  veh.wt,
  veh.spd = NULL,
  veh.type = NULL,
  em.type = c("pm2.5", "pm10"),
  em.source = c("brake", "tyre", "road", "resusp"),
  brk.regen = FALSE,
  route.def = c("urban", "rural", "motorway"),
  route.source = "uk naei",
  verbose = FALSE,
  ...
)

ef_embrs1_brake_pm2.5(em.type, em.source, ...)

ef_embrs1_tyre_pm2.5(em.type, em.source, ...)

ef_embrs1_road_pm2.5(em.type, em.source, ...)

ef_embrs1_resusp_pm2.5(em.type, em.source, ...)

ef_embrs1_brake_pm10(em.type, em.source, ...)

ef_embrs1_tyre_pm10(em.type, em.source, ...)

ef_embrs1_road_pm10(em.type, em.source, ...)

ef_embrs1_resusp_pm10(em.type, em.source, ...)

ef_embrs2_brake_pm(
  veh.wt,
  veh.spd = NULL,
  veh.type = NULL,
  brk_b = NULL,
  em.type = c("pm2.5", "pm10"),
  em.source = c("brake"),
  brk.regen = FALSE,
  route.def = c("urban", "rural", "motorway"),
  route.source = "uk naei",
  verbose = FALSE,
  ...
)

ef_embrs2_brake_pm2.5(em.type, em.source, ...)

ef_embrs2_brake_pm10(em.type, em.source, ...)

ef_embrs2_tyre_pm(
  veh.wt,
  veh.spd = NULL,
  veh.type = NULL,
  tyr_t = NULL,
  em.type = c("pm2.5", "pm10"),
  em.source = c("tyre"),
  brk.regen = FALSE,
  route.def = c("urban", "rural", "motorway"),
  route.source = "uk naei",
  verbose = FALSE,
  ...
)

ef_embrs2_tyre_pm2.5(em.type, em.source, ...)

ef_embrs2_tyre_pm10(em.type, em.source, ...)
}
\arguments{
\item{veh.wt}{(numeric, required) vehicle weight (in kg).}

\item{veh.spd}{(numeric, required) vehicle speed (in km/hr).}

\item{veh.type}{(character) type of vehicle, e.g bus.}

\item{em.type}{(character) type of emissions to predict, by default PM2.5
and PM10.}

\item{em.source}{(character) emission source, by default brake, tyre, road
and resuspension, but can be any combination.}

\item{brk.regen}{vehicle regenerative braking, default \code{FALSE} means
not installed/used or a numeric (0-1) for efficiency if installed/used,
e.g. 0.25 = 25\% efficient (equivalent to a 75\% of a conventional
non-regenerative brake contribution).}

\item{route.def}{(character) route definition(s).}

\item{route.source}{(character) route source(s).}

\item{verbose}{(logical) If TRUE, include methods details
when reporting EF predictions.}

\item{...}{other arguments, often passed on.}

\item{brk_b}{(numeric; ef_embrs2_brake... only) brake work parameter
proposed in Tivey et al (2023).}

\item{tyr_t}{(numeric; ef_embrs2_brake... only) tyre work parameter
proposed in Tivey et al (2023).}
}
\value{
These functions build data.frames of vehicle weight and speed,
(and brake and tyre work) based emission factors for non-exhaust PM2.5
and PM10. \code{ef_embrs1_nee_pm} is the main (weight and speed)
function, other \code{ef_embrs1...} functions are wrappers for this for
single source and type emissions factors, and \code{ef_embrs2...} functions
are brake and tyre work based functions proposed in Tivey et al (2023).
}
\description{
Functions to estimated vehicle weight, speed and brake-work-based
non-exhaust PM2.5 and PM10 emission factors based the extension of Beddows
proposed in embrs paper.
}
\note{
These may be moving to vein at some point...
}
\references{
These functions are based on methods developed and reported by:

Beddows, D.C. and Harrison, R.M., 2021. PM10 and PM2.5 emission factors for
non-exhaust particles from road vehicles: Dependence upon vehicle mass and
implications for battery electric vehicles. Atmospheric Environment, 244,
p.117886. \url{https://doi.org/10.1016/j.atmosenv.2020.117886}.

And, extrapolated to speed, brake and tyre-wear related emission factors in:

Tivey, J., Davies, H.C., Levine, J.G., Zietsman, J., Bartington, S.,
Ibarra-Espinosa, S. and Ropkins, K, 2023. Meta-Analysis as Early Evidence on
the Particulate Emissions Impact of EURO VI on Battery Electric Bus Fleet
Transitions. Sustainability 15, 1522. \url{https://doi.org/10.3390/su15021522}
}
