% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_vein_eea_exhaust.R
\name{ef_vein_eea_exhaust}
\alias{ef_vein_eea_exhaust}
\alias{ef_vein_eea_exhaust_pm2.5}
\alias{ef_vein_eea_exhaust_pm10}
\title{ef_vein_eea_exhaust}
\source{
Average speed curves are based on European Environment Agency 2019 EMEP/EEA
methods as documented in:

European Environment Agency, 2019. EMEP/EEA air pollutant
emission inventory guidebook.
https://www.eea.europa.eu/publications/emep-eea-guidebook-2019

And as implemented in R package vein:

Ibarra-Espinosa, S., Ynoue, R., O'Sullivan, S., Pebesma, E., Andrade, M. D. F.,
and Osses, M.: VEIN v0.2.2: an R package for bottom-up vehicular emissions
inventories, Geosci. Model Dev., 11, 2209-2229,
https://doi.org/10.5194/gmd-11-2209-2018, 2018.

Ibarra-Espinosa S (2022). \emph{vein: Vehicular Emissions Inventories}. R package
version 0.9.9, \url{https://CRAN.R-project.org/package=vein}.

Fuel corrections are based on:

TRL Emission Factors - scaling factors for mileage and improvements in fuel quality
https://www.gov.uk/government/publications/road-vehicle-emission-factors-2009
}
\usage{
ef_vein_eea_exhaust(
  veh.spd = NULL,
  veh.type = NULL,
  veh.wt = NULL,
  veh.segment = NULL,
  em.type = NULL,
  route.def = NULL,
  eng.type = NULL,
  eng.fuel = NULL,
  fuel.corr = TRUE,
  em.source = "exhaust",
  euro.class = NULL,
  exh.tech = NULL,
  eng.load = NULL,
  route.slope = NULL,
  route.mode = NULL,
  verbose = FALSE,
  ...
)

ef_vein_eea_exhaust_pm2.5(em.type = "pm2.5", ...)

ef_vein_eea_exhaust_pm10(em.type = "pm10", ...)
}
\arguments{
\item{veh.spd}{average vehicle speed, in km/h.}

\item{veh.type}{vehicle type, currently only coded for buses, see Note.}

\item{veh.wt}{vehicle weight, in km.}

\item{veh.segment}{EEA vehicle sub-classification, specific to vehicle type.
For buses, embrs assigns this using veh.type and veh.wt, so it can often be
ignored.}

\item{em.type}{type of emission, currently PM, PM2.5 or PM10, see Note.}

\item{route.def}{route description, optional for ef_vein_eea() functions.}

\item{eng.type}{engine type: ice, etc.}

\item{eng.fuel}{fuel used by the vehicle: diesel, biodiesel, cng, etc.}

\item{fuel.corr}{(logical) apply fuel correction, see Note.}

\item{em.source}{emission source: currently, exhaust only.}

\item{euro.class}{EURO classification: PRE, I, II, III, IV, V,
V or VI only.}

\item{exh.tech}{exhaust technology: for euro.class IV SCR or EGR,
for V  SCR or EGR, and for VI DPF+SCR only}

\item{eng.load}{engine load, as proportion: 0. 0.5 or 1 only.}

\item{route.slope}{route slope, as proportion: -0.06, -0.04, -0.02, 0,
0.02, 0.04 or 0.06 only.}

\item{route.mode}{route driving/journey mode: Urban Peak, Urban Off Peak,
Rural, Motorway, or NA (not defined).}

\item{verbose}{(logical) If TRUE, include methods details
when reporting EF predictions.}

\item{...}{other arguments, currently ignored}
}
\description{
Functions to calculate vehicle speed-based hot exhaust
emission factors using on EMEP/EEA guidelines 2019 methods.
}
\note{
The embrs::ef_vein_eea functions are currently only coded for
buses, \code{veh.type="bus"}.

Strictly, EEA 2019 PM exhaust emission factors are PM, but because
vehicle exhaust PM is fine it is typically assumed that:

PM10 = PM2.5 = PM for exhaust emissions from most conventional diesel and
gasoline vehicles.

Fuel Corrections are applied to account for improvements in emissions
associated with improvements in fuel quality,
}
