% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.em.class.R
\name{em.class}
\alias{em.class}
\alias{em_test}
\alias{+.embrs_act}
\alias{*.embrs_acts}
\alias{*.embrs_veh}
\title{em.class}
\usage{
em_test(...)

\method{+}{embrs_act}(x, y, ...)

\method{*}{embrs_acts}(x, y, ...)

\method{*}{embrs_veh}(x, y, ...)
}
\value{
These functions make em class objects, which...
}
\description{
Functions to build emissions models using embrs objects.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and Beddows references...
}
