% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.generics.R
\name{embrs.generics}
\alias{embrs.generics}
\alias{+.embrs}
\alias{as.embrs_fleet.embrs_vehicle}
\alias{*.embrs}
\alias{print.embrs}
\alias{plot.embrs}
\alias{print.embrs_vehicle}
\title{embrs generics}
\usage{
\method{+}{embrs}(x, y, ...)

\method{*}{embrs}(x, y, ...)

\method{print}{embrs}(x, ...)

\method{plot}{embrs}(x, ...)

\method{print}{embrs_vehicle}(x, ...)
}
\arguments{
\item{x, y}{function arguments}
}
\value{
These functions make vehicle and fleet class embrs objects.
}
\description{
Functions to build vehicle, fleet and route objects for use in
embrs emission models.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and Beddows references...
}
