% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.vehicles.R
\name{embrs.vehicles}
\alias{embrs.vehicles}
\alias{embrs_ice}
\alias{embrs_bev}
\title{common vehicle objects}
\usage{
embrs_ice(
  veh.wt = NULL,
  veh.type = NULL,
  eng.fuel = NULL,
  euro.class = NULL,
  n = 1,
  method = "embrs1",
  name = NULL,
  ...
)

embrs_bev(
  veh.wt = NULL,
  veh.type = NULL,
  n = 1,
  method = "embrs1",
  name = NULL,
  ...
)
}
\arguments{
\item{veh.wt}{(required numeric) weight of vehicle in kg.}

\item{veh.type}{(required character) type of vehicle, e.g bus.}

\item{eng.fuel}{(required character) the engine fuel used by the vehicle,
currently only diesel.}

\item{euro.class}{(required character) vehicle EURO classification,
e.g. PRE, I, II, III, IV, V or VI.}

\item{n}{(numeric) number of vehicles, default 1.}

\item{method}{(character) vehicle model, currently one of: beddows,
embrs1 (default) or embrs2.}

\item{name}{(character) name of the vehicle object.}

\item{...}{other arguments, currently passed on to emission factor (ef)
models. In some cases, these may request additional inputs, e.g. when
ef_vein_eea has more than one ef model for a given
vehicle/engine/fuel/euro.class combination exhaust emission model it may
also request the exhaust technology (exh.tech), see ?ef_vein_eea_exhaust.}
}
\value{
These functions make common vehicle objects for use in \strong{embrs}
emissions models. The main models ice_build and bev_build are for internal
compression engine and battery electric vehicles, respectively.
}
\description{
Common builds for \strong{embrs} vehicle builds.
}
\note{
Work in progress; these may change...
}
\references{
These functions are based on methods developed and reported by:

Beddows, D.C. and Harrison, R.M., 2021. PM10 and PM2.5 emission factors for
non-exhaust particles from road vehicles: Dependence upon vehicle mass and
implications for battery electric vehicles. Atmospheric Environment, 244,
p.117886. \url{https://doi.org/10.1016/j.atmosenv.2020.117886}.

And, extrapolated to speed (embrs1) and, break/tyre-wear (embrs2)
related emission factors in:

Tivey, J., Davies, H.C., Levine, J.G., Zietsman, J., Bartington, S.,
Ibarra-Espinosa, S. and Ropkins, K, 2023. Meta-Analysis as Early Evidence on
the Particulate Emissions Impact of EURO VI on Battery Electric Bus Fleet
Transitions. Sustainability 15, 1522. \url{https://doi.org/10.3390/su15021522}
}
