% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs-package.R
\docType{package}
\name{embrs}
\alias{embrs}
\title{What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
This package was splat packed.
}
\section{package functions}{

\link{to be added}
}

