% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_misc.R
\name{misc.mod}
\alias{misc.mod}
\alias{naei_route2spd}
\title{misc modifiers}
\usage{
naei_route2spd(veh.type = NULL, route.def = NULL, route.source = NULL, ...)
}
\arguments{
\item{veh.type}{(character) vehicle type e.g. bus}

\item{route.source}{(character) route source, for NAEI}

\item{...}{other arguments, often passed on.}

\item{veh.def}{(character) route definition, for NAEI urban, rural or
motorway}
}
\value{
These estimated veh speed for the supplied route.
}
\description{
functions that currently don't have better homes...
}
\note{
These are currently the estimates used by \strong{embrs} models...
}
\references{
naei_route2spd is based on speed estimates in:

Brown, P., Wakeling, D., Pang, Y. and Murrells, T. Methodology for the UK's
road transport emissions inventory: Version for the 2016 National Atmospheric
Emissions Inventory. Report for the Department for Business, Energy &
Industrial Strategy. 2018.
https://uk-air.defra.gov.uk/assets/documents/reports/cat07/1804121004_Road_transport_emissions_methodology_report_2018_v1.1.pdf
}
