% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.activity.class.R, R/embrs.vehicle.class.R
\name{print.embrs_act}
\alias{print.embrs_act}
\alias{print.embrs_acts}
\alias{+.embrs_act}
\alias{+.embrs_acts}
\alias{vehicle.class}
\alias{bus_ice}
\alias{bus_bev}
\alias{print.embrs_veh}
\alias{print.embrs_fleet}
\title{vehicle class}
\usage{
\method{print}{embrs_act}(x, ...)

\method{print}{embrs_acts}(x, ...)

\method{+}{embrs_act}(x, y, ...)

\method{+}{embrs_acts}(x, y, ...)

bus_ice(veh.wt, n = 1, ...)

\method{print}{embrs_veh}(x, ...)

\method{print}{embrs_fleet}(x, ...)
}
\arguments{
\item{veh.wt}{(numeric) number of vehicles, by default 1.}
}
\value{
These functions make vehicle class objects, which .
}
\description{
Functions to build vehicle objects for use in embrs emission
models.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and Beddows references...
}
