% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus.objects.R
\name{bus.objects}
\alias{bus.objects}
\alias{bus_ice}
\alias{bus_bev}
\title{bus objects}
\usage{
bus_ice(
  veh.wt = NULL,
  eng.fuel = NULL,
  euro.class = NULL,
  ...,
  veh.type = "bus"
)

bus_bev(veh.wt = NULL, ..., veh.type = "bus")
}
\arguments{
\item{veh.wt}{(required numeric) weight of vehicle in kg.}

\item{eng.fuel}{(required character) the engine fuel used by the vehicle,
e.g. diesel.}

\item{euro.class}{(required character) vehicle EURO classification,
e.g. PRE, I, II, III, IV, V or VI.}

\item{...}{other arguments, passed on to common vehicle build functions.
Arguments depend on the vehicle/engine/fuel combination. See embrs.vehicles
help for further details.}

\item{veh.type}{(character) type of vehicle, default bus.}
}
\value{
These functions make bus models for \strong{embrs} emissions
models. The main models bus_ice and bus_bev are internal compression engine
and battery electric buses, respectively.
}
\description{
Bus objects for use in \strong{embrs} models.
}
\note{
Work in progress...
}
\references{
These functions are based on methods developed and reported by:

Beddows, D.C. and Harrison, R.M., 2021. PM10 and PM2.5 emission factors for
non-exhaust particles from road vehicles: Dependence upon vehicle mass and
implications for battery electric vehicles. Atmospheric Environment, 244,
p.117886. \url{https://doi.org/10.1016/j.atmosenv.2020.117886}.

And, extrapolated to speed and, break/tyre-wear related emission factors in:

Tivey, J., Davies, H.C., Levine, J.G., Zietsman, J., Bartington, S.,
Ibarra-Espinosa, S. and Ropkins, K, 2023. Meta-Analysis as Early Evidence on
the Particulate Emissions Impact of EURO VI on Battery Electric Bus Fleet
Transitions. Sustainability 15, 1522. \url{https://doi.org/10.3390/su15021522}
}
