% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus.objects.R
\name{bus.objects}
\alias{bus.objects}
\alias{embrs_vehicle}
\alias{bus_ice}
\alias{bus_bev}
\alias{bus_ice_beddows}
\alias{bus_bev_beddows}
\alias{bus_ice_embrs1}
\alias{bus_bev_embrs1}
\alias{bus_ice_embrs2}
\alias{bus_bev_embrs2}
\title{bus objects}
\usage{
embrs_vehicle(name = NULL, x = NULL, ...)

bus_bev(...)

bus_ice(...)

bus_ice_beddows(veh.wt, euro.class = NULL, n = 1, name = NULL, ...)

bus_bev_beddows(veh.wt, n = 1, name = NULL, ...)

bus_ice_embrs1(veh.wt, euro.class = NULL, n = 1, name = NULL, ...)

bus_bev_embrs1(veh.wt, n = 1, name = NULL, ...)

bus_ice_embrs2(veh.wt, euro.class = NULL, n = 1, name = NULL, ...)

bus_bev_embrs2(veh.wt, n = 1, name = NULL, ...)
}
\arguments{
\item{name}{(character) name of the vehicle object.}

\item{x}{(for \code{embrs_vehicle}) a list of args to
using when building an \strong{embrs} vehicle/fleet object.}

\item{...}{other arguments, currently ignored}

\item{veh.wt}{(required numeric) weight of vehicle in kg.}

\item{euro.class}{(required character) vehicle EURO classification,
can only be PRE, I, II, III, IV, V+EGR, V+SCR or VI.}

\item{n}{(numeric) number of vehicles, defualt 1.}
}
\description{
Bus objects for use in \strong{embrs} models.
}
\note{
These may be moving to vein at some point...
}
\references{
\link{add embrs and Beddows references...}
}
