% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus.objects.R
\name{bus.objects}
\alias{bus.objects}
\alias{bus_ice}
\alias{bus_bev}
\alias{bus_ice_beddows}
\alias{bus_bev_beddows}
\alias{bus_ice_embrs1}
\alias{bus_bev_embrs1}
\alias{bus_ice_embrs2}
\alias{bus_bev_embrs2}
\title{bus objects}
\usage{
bus_ice(
  veh.wt = NULL,
  eng.fuel = NULL,
  euro.class = NULL,
  n = 1,
  name = NULL,
  ...
)

bus_bev(veh.wt = NULL, n = 1, name = NULL, ...)

bus_ice_beddows(...)

bus_ice_embrs1(...)

bus_ice_embrs2(...)

bus_bev_beddows(...)

bus_bev_embrs1(...)

bus_bev_embrs2(...)
}
\arguments{
\item{veh.wt}{(required numeric) weight of vehicle in kg.}

\item{eng.fuel}{(required character) the fuel used by the bus, currently
only diesel.}

\item{euro.class}{(required character) vehicle EURO classification,
e.g. PRE, I, II, III, IV, V or VI.}

\item{n}{(numeric) number of vehicles, default 1.}

\item{name}{(character) name of the vehicle object.}

\item{...}{other arguments, currently ignored}
}
\value{
These functions make bus model for use in \strong{embrs} emissions
models. The main models bus_ice and bus_bev are internal compression engine
and battery electric buses, respectively.
}
\description{
Bus objects for use in \strong{embrs} models.
}
\note{
Work in progress...
}
\references{
These functions are based on methods developed and reported by:

Beddows, D.C. and Harrison, R.M., 2021. PM10 and PM2.5 emission factors for
non-exhaust particles from road vehicles: Dependence upon vehicle mass and
implications for battery electric vehicles. Atmospheric Environment, 244,
p.117886. \url{https://doi.org/10.1016/j.atmosenv.2020.117886}.

And, extrapolated to speed and, break/tyre-wear related emission factors in:

Tivey, J., Davies, H.C., Levine, J.G., Zietsman, J., Bartington, S.,
Ibarra-Espinosa, S. and Ropkins, K, 2023. Meta-Analysis as Early Evidence on
the Particulate Emissions Impact of EURO VI on Battery Electric Bus Fleet
Transitions. Sustainability 15, 1522. \url{https://doi.org/10.3390/su15021522}
}
