% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_beddows.R
\name{ef_beddows}
\alias{ef_beddows}
\alias{ef_beddows_nee_pm}
\alias{ef_beddows_brake_pm2.5}
\alias{ef_beddows_tyre_pm2.5}
\alias{ef_beddows_road_pm2.5}
\alias{ef_beddows_resusp_pm2.5}
\alias{ef_beddows_brake_pm10}
\alias{ef_beddows_tyre_pm10}
\alias{ef_beddows_road_pm10}
\alias{ef_beddows_resusp_pm10}
\title{Beddows Non-Exhaust Emissions Factors}
\usage{
ef_beddows_nee_pm(
  veh.wt,
  em.type = c("pm2.5", "pm10"),
  em.source = c("brake", "tyre", "road", "resusp"),
  brk.regen = FALSE,
  route.def = c("urban", "rural", "motorway"),
  route.source = "uk naei",
  verbose = FALSE,
  ...
)

ef_beddows_brake_pm2.5(em.type, em.source, ...)

ef_beddows_brake_pm10(em.type, em.source, ...)

ef_beddows_tyre_pm2.5(em.type, em.source, ...)

ef_beddows_tyre_pm10(em.type, em.source, ...)

ef_beddows_road_pm2.5(em.type, em.source, ...)

ef_beddows_road_pm10(em.type, em.source, ...)

ef_beddows_resusp_pm2.5(em.type, em.source, ...)

ef_beddows_resusp_pm10(em.type, em.source, ...)
}
\arguments{
\item{veh.wt}{(numeric, required) vehicle weight (in kg).}

\item{em.type}{(character) type of emissions to predict, by default PM2.5
and PM10.}

\item{em.source}{(character) emission source, by default brake, tyre, road
and resuspension, but can be any combination.}

\item{brk.regen}{vehicle regenerative braking, default \code{FALSE} means
not installed/used or a numeric (0-1) for efficiency if installed/used,
e.g. 0.25 = 25\% efficient (equivalent to a 75\% of a conventional
non-regenerative brake contribution).}

\item{route.def}{(character) route definitions, by default urban, rural and motorway.}

\item{route.source}{(character) route sources, must be UK NAEI.}

\item{...}{other arguments, often passed on.}
}
\value{
These functions build data.frames of urban, rural and motorway emission
factors for non-exhaust PM2.5 and PM10. \code{ef_beddows_nee_pm} is the main
function, and others are wrappers for single soucre and type emissions factors.
}
\description{
Functions to estimated vehicle weight-based non-exhaust
PM2.5 and PM10 emission factors based on methods of Beddows.
}
\note{
These may be moving to vein at some point...
}
\references{
These functions are based on methods developed and reported by:

Beddows, D.C. and Harrison, R.M., 2021. PM10 and PM2. 5 emission factors for
non-exhaust particles from road vehicles: Dependence upon vehicle mass and
implications for battery electric vehicles. Atmospheric Environment, 244,
p.117886. \url{https://doi.org/10.1016/j.atmosenv.2020.117886}
}
