% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.generics.R
\name{embrs.generics}
\alias{embrs.generics}
\alias{+.embrs}
\alias{as.embrs_fleet.embrs_vehicle}
\alias{*.embrs}
\alias{print.embrs}
\alias{plot.embrs}
\alias{print.embrs_vehicle}
\alias{print.embrs_route}
\alias{print.embrs_model}
\title{embrs generics}
\usage{
\method{+}{embrs}(x, y, ...)

\method{*}{embrs}(x, y, ...)

\method{print}{embrs}(x, ...)

\method{plot}{embrs}(x, plot.type = "by.em", em.type = "all", ...)

\method{print}{embrs_vehicle}(x, ...)

\method{print}{embrs_route}(x, ...)

\method{print}{embrs_model}(x, ...)
}
\arguments{
\item{x, y}{function arguments}

\item{plot.type}{(for plot only, character) the type of plot to generate,
options: 'by.em' standard plot; 'by.vehicle'.
See Note.
\code{NULL} to disable/show all emissions types calculated.}

\item{em.type}{(for plot only, character) the emissions to generate plot
for, options: 'all' for all emissions types, or 'just.pm' for just
particulates. See Note.}
}
\value{
These functions make vehicle and fleet class \strong{embrs} objects.
}
\description{
Functions to build vehicle, fleet and route objects for use in
embrs emission models.
}
\note{
These functions are in development, so please be aware that options
outputs may change significantly.
plot.type = 'by.vehicle' and em.type = 'just.pm' produces plots like those
used in Tivey el al (2023) to compare particulate emissions from different
buses types.
}
\references{
Tivey, J., Davies, H.C., Levine, J.G., Zietsman, J., Bartington, S.,
Ibarra-Espinosa, S. and Ropkins, K, 2023. Meta-Analysis as Early Evidence on
the Particulate Emissions Impact of EURO VI on Battery Electric Bus Fleet
Transitions. Sustainability 15, 1522. \url{https://doi.org/10.3390/su15021522}
}
