% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs.generics.R
\name{embrs.generics}
\alias{embrs.generics}
\alias{+.embrs}
\alias{as.embrs_fleet.embrs_vehicle}
\alias{*.embrs}
\alias{print.embrs}
\alias{plot.embrs}
\alias{print.embrs_vehicle}
\alias{print.embrs_route}
\alias{print.embrs_model}
\title{embrs generics}
\usage{
\method{+}{embrs}(x, y, ...)

\method{*}{embrs}(x, y, ...)

\method{print}{embrs}(x, ...)

\method{plot}{embrs}(x, em.type = "just.pm", ...)

\method{print}{embrs_vehicle}(x, ...)

\method{print}{embrs_route}(x, ...)

\method{print}{embrs_model}(x, ...)
}
\arguments{
\item{x, y}{function arguments}

\item{em.type}{(for plot only, character) the emissions types to calculate
emission factors for, special cases: 'just.pm' just PM emissions types;
\code{NULL} to disable/show all emissions types calculated.}
}
\value{
These functions make vehicle and fleet class \strong{embrs} objects.
}
\description{
Functions to build vehicle, fleet and route objects for use in
embrs emission models.
}
\note{
These functions are in development, so please be aware that options
outputs may change significantly.
}
\references{
add embrs and Beddows references...
}
