% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrs-package.R
\docType{package}
\name{embrs}
\alias{embrs}
\title{Emission Model Builder & Reporter Syntax}
\description{
(in development) \strong{embrs} package.
}
\details{
This package was splat packed.
}
\section{package functions}{

\link{to be added}
}

