% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.objects.R
\name{route.objects}
\alias{route.objects}
\alias{embrs_route}
\alias{route_naei_rural}
\alias{route_naei_urban}
\alias{route_naei_motorway}
\alias{route_naei_urm}
\alias{route_ukbc_outerlondon}
\alias{route_ukbc_innerlondon}
\alias{route_ukbc_rural}
\alias{route_ukbc_oir}
\title{route.objects}
\usage{
embrs_route(
  name = NULL,
  route.def = NULL,
  route.source = NULL,
  route.dist = 1,
  route.slope = 0,
  ...
)

route_naei_rural(
  name = NULL,
  route.def = "rural",
  route.source = "uk naei",
  veh.spd = naei_route2spd,
  ...
)

route_naei_urban(
  name = NULL,
  route.def = "urban",
  route.source = "uk naei",
  veh.spd = naei_route2spd,
  ...
)

route_naei_motorway(
  name = NULL,
  route.def = "motorway",
  route.source = "uk naei",
  veh.spd = naei_route2spd,
  ...
)

route_naei_urm(...)

route_ukbc_outerlondon(
  name = NULL,
  route.def = "outer london",
  route.source = "ukbc",
  veh.spd = ukbc_route2spd,
  brk_b = ukbc_route2brk_b,
  tyr_t = ukbc_route2tyr_t,
  ...
)

route_ukbc_innerlondon(
  name = NULL,
  route.def = "inner london",
  route.source = "ukbc",
  veh.spd = ukbc_route2spd,
  brk_b = ukbc_route2brk_b,
  tyr_t = ukbc_route2tyr_t,
  ...
)

route_ukbc_rural(
  name = NULL,
  route.def = "rural",
  route.source = "ukbc",
  veh.spd = ukbc_route2spd,
  brk_b = ukbc_route2brk_b,
  tyr_t = ukbc_route2tyr_t,
  ...
)

route_ukbc_oir(...)
}
\arguments{
\item{name}{(character) optional object name, by default the route source and type.}

\item{route.def}{(character) required route description.}

\item{route.source}{(character) required source of route model.}

\item{route.dist}{(numeric) route or journey distance in km, default 1.}

\item{route.slope}{(numeric) route or journey slope, default 0.}

\item{...}{other arguments, currently ignored.}

\item{veh.spd}{(numeric or function) the vehicle speed in km/hr.}
}
\description{
Route objects for use in embrs models.
}
\note{
trying to streamline these, so likely to be subject to change
}
