% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.objects.R
\name{route.objects}
\alias{route.objects}
\alias{route_naei_rural}
\alias{route_naei_urban}
\alias{route_naei_motorway}
\alias{route_naei_rum}
\title{route.objects}
\usage{
route_naei_rural(name = NULL, ...)

route_naei_urban(name = NULL, ...)

route_naei_motorway(name = NULL, ...)

route_naei_rum()
}
\arguments{
\item{...}{other arguments, currently ignored}
}
\description{
Route objects for use in embrs models.
}
\note{
These may be moving to vein at some point...
}
\references{
add embrs and naei references...
}
